requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
chromadb
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langchain_core
pdfplumber
networkx
pyvis
tqdm
sentence-transformers